requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.3.8,<2
pillow==10.1.0
pytesseract==0.3.10
aiofiles==23.2.0
slowapi==0.1.9
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
structlog==23.2.0
prometheus-client==0.19.0
